+++ packages/_private/e2e-demo-test/src/AllComponents/SynSelect.spec.ts
@@ -203,5 +203,47 @@ test.describe('<SynSelect />', () => {
         expect(secondValue).toEqual('Option 2');
       });
     }); // regression#1036
+
+    test.describe(`Regression#1056: ${name}`, () => {
+      test('should update pre selected value correct for async changed delimiter', async ({ page }) => {
+        const AllComponents = new AllComponentsPage(page, port);
+        await AllComponents.loadInitialPage();
+        await AllComponents.activateItem('selectLink');
+
+        await expect(AllComponents.getLocator('selectContent')).toBeVisible();
+
+        const select = await AllComponents.getLocator('select1056DelimiterPreValue');
+
+        const firstValue = await select.evaluate((ele: SynSelect) => ele.value);
+        expect(firstValue).toEqual('');
+
+        await select.evaluate((ele: SynSelect) => {
+          ele.delimiter = '|';
+        });
+
+        const secondValue = await select.evaluate((ele: SynSelect) => ele.value);
+        expect(secondValue).toEqual('Option 2');
+      });
+
+      test('should update async pre selected value correct for async changed delimiter', async ({ page }) => {
+        const AllComponents = new AllComponentsPage(page, port);
+        await AllComponents.loadInitialPage();
+        await AllComponents.activateItem('selectLink');
+
+        await expect(AllComponents.getLocator('selectContent')).toBeVisible();
+
+        const select = await AllComponents.getLocator('select1056DelimiterAsyncPreValue');
+
+        const firstValue = await select.evaluate((ele: SynSelect) => ele.value);
+        expect(firstValue).toEqual('');
+
+        await select.evaluate((ele: SynSelect) => {
+          ele.delimiter = '|';
+        });
+
+        const secondValue = await select.evaluate((ele: SynSelect) => ele.value);
+        expect(secondValue).toEqual('Option 2');
+      });
+    }); // regression#1056
   }); // End frameworks
 }); // </syn-select>